Refactor code structure: move types and configurations to dedicated modules, enhance error handling, and update imports
