+++ src/main.rs
@@ -1,48 +1,7 @@
-use crate::state::EmailSender;
-use tokio::net::{TcpStream};
-use std::net::SocketAddr;
-use tokio::time;
-
-
 mod handler;
-mod types;
-mod state;
+mod models;
 mod errors;
-
-
-const MAX_TIMEOUT_SECS: u64 = 180; // 3 minutes
-
-
-/// Handle a single client connection. This function is spawned as a new task for each connection.
-async fn handle_connection(socket: TcpStream, addr: SocketAddr, amqp_tx: EmailSender) {
-    // Create a new UUID for the connection/session
-    let conn_id = uuid::Uuid::new_v4();
-
-    // Create a new email handler
-    let client = handler::email::EmailHandler::new(socket, conn_id);
-
-    log::debug!("[conn={}] Handling connection from: {}", conn_id, addr);
-
-    // Run the client with a timeout
-    match time::timeout(time::Duration::from_secs(MAX_TIMEOUT_SECS), client.run()).await {
-        Ok(Ok(email)) => {
-            log::info!("[conn={}] Received email: {}", conn_id, email.message_id);
-
-            // Send the email to the AMQP channel
-            if let Err(e) = amqp_tx.send(email).await {
-                log::error!("[conn={}] Failed to send email to AMQP channel: {}", conn_id, e);
-            }
-        }
-
-        Ok(Err(e)) => {
-            log::error!("[conn={}] Error handling client {}: {}", conn_id, addr, e);
-        }
-
-        Err(_) => {
-            log::warn!("[conn={}] Connection handler timed out after {} seconds for client: {}", conn_id, MAX_TIMEOUT_SECS, addr);
-        }
-    }
-}
+mod state;
 
 
 #[tokio::main]
@@ -61,7 +20,7 @@ async fn main() -> Result<(), errors::LSMTPError> {
 
         // Spawn a new task to handle the client connection
         tokio::spawn(async move {
-            handle_connection(socket, addr, amqp_tx).await;
+            state::handle_connection(socket, addr, amqp_tx).await;
         });
     }
 }
