+++ src/models/mod.rs
@@ -1,0 +1,2 @@
+pub mod configs;
+pub mod email;
